Ajustando .gitignore e removendo arquivos gerados(target, classes, test-classes..), adicionando DTO'S/Exceptions basicas, validação basica para regra de negocio, controller autor inicial basico feito, correcao de repo local com repo remoto

+++ src/main/java/com/erickgx/libraryapi/service/AutorService.java
@@ -1,7 +1,10 @@
 package com.erickgx.libraryapi.service;
 
+import com.erickgx.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import com.erickgx.libraryapi.models.Autor;
 import com.erickgx.libraryapi.repository.AutorRepository;
+import com.erickgx.libraryapi.repository.LivroRepository;
+import com.erickgx.libraryapi.validator.AutorValidator;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,12 +15,17 @@ import java.util.UUID;
 public class AutorService {
 
     private final AutorRepository repository;
+    private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
-    public AutorService(AutorRepository repository){
+    public AutorService(AutorRepository repository, AutorValidator validator, LivroRepository livroRepository) {
         this.repository = repository;
+        this.validator = validator;
+        this.livroRepository = livroRepository;
     }
 
     public Autor salvar (Autor autor){
+         validator.validar(autor);
         return repository.save(autor);
     }
 
@@ -26,6 +34,10 @@ public class AutorService {
     }
 
     public void deletar(Autor autor){
+        if (possuiLivro(autor)){
+            throw new OperacaoNaoPermitidaException(
+                    "Não é permitido excluir um Autor que possui livros cadastrados!");
+        }
         repository.delete(autor);
     }
 
@@ -43,5 +55,17 @@ public class AutorService {
         }
 
         return repository.findAll();
+    }
+
+    public void atualizar(Autor autor){
+        if (autor.getId() == null){
+            throw new IllegalArgumentException("Para atualizar é necessario que o autor já esteja salvo na base");
+        }
+        validator.validar(autor);
+        repository.save(autor);
+    }
+
+    public boolean possuiLivro(Autor autor){
+        return livroRepository.existsByAutor(autor);
     }
 }
